Convert README from markdown to asciidoc
I've recently discovered asciidoc, an alternative to markdown that has a
lot of cool features and a similar ascii-based gets-out-of-your-way
syntax. See http://asciidoctor.org/docs/what-is-asciidoc/ and
http://asciidoctor.org/docs/user-manual/#compared-to-markdown for a nice
intro, and http://asciidoctor.org/docs/asciidoc-writers-guide/ for a
deep dive into the format.

For instance, with asciidoc you can get a table of contents without
needing an external program to generate and hardcode it; you can also
get automatic section counters (and other kinds of counters) without
having to hardcode numbers on your document.

So I basically wanted an excuse to play with it. Plus, GitHub can
already display it, as it does for markdown, so win-win!